+++ broker/app.py
@@ -14,12 +14,12 @@ from sap.cf_logging import flask_logging
 # `flask db migrate`
 from broker import models  # noqa: F401
 from broker.api import API, ClientError
-from broker.commands.waf import add_dedicated_alb_waf_web_acls
-from broker.extensions import config, db, migrate
-from broker.duplicate_certs import (
+from broker.commands.duplicate_certs import (
     log_duplicate_alb_cert_metrics,
     remove_duplicate_alb_certs,
 )
+from broker.commands.waf import add_dedicated_alb_waf_web_acls
+from broker.extensions import config, db, migrate
 from broker.errors import handle_exception
